app/requirements.txt
PyYAML>=6.0.1
pandas>=2.0.0
kafka-python>=2.0.2
